Move the hook into the hooks folder
