+++ src/components/SignupCTA/index.tsx
@@ -1,8 +1,7 @@
 import React from 'react'
 import { CallToAction } from 'components/CallToAction'
-import { useValues } from 'kea'
-import { posthogAnalyticsLogic } from 'logic/posthogAnalyticsLogic'
 import { RenderInClient } from 'components/RenderInClient'
+import usePostHog from '../../hooks/usePostHog'
 
 /**
  * A signup CTA that directs to the correct region (EU or US) based on feature flag.
@@ -20,7 +19,7 @@ export const SignupCTA = ({
     width?: string
     event?: any
 }): JSX.Element => {
-    const { posthog } = useValues(posthogAnalyticsLogic)
+    const posthog = usePostHog()
 
     return (
         <RenderInClient
